+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 - ILIAS 5.4 support
 - Drop ILIAS 5.1 and ILIAS 5.2 support
 - Implement uninstaller
+- Fix cancel object mapping button
+- Fix no array pass to `array_values`